requirements.txt
astroid==1.4.8
backports.functools-lru-cache==1.2.1
cocos2d==0.6.4
configparser==3.5.0
decorator==4.0.10
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
networkx==1.11
numpy==1.11.1
Pillow==3.3.1
pyglet==1.2.4
pylint==1.6.4
PyYAML==3.11
scipy==0.18.0
six==1.10.0
wrapt==1.10.8
